Moved existing `tree/` modules to `internalTree/`
